requirements.txt
fastapi==0.110.0

langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.1.7

qdrant-client==1.17.0

sentence-transformers==2.6.1
transformers==4.37.2
huggingface_hub==0.20.3

atlassian-python-api==3.41.10
beautifulsoup4==4.12.3

fastmcp==2.13.3
